+++ agentid/__init__.py
@@ -1,3 +1,3 @@
 """AgentAction package."""
 
-__version__ = "0.36.1rc1"
+__version__ = "0.37.0rc1"